+++ src/components/App/Renderer/MainMenu/index.js
@@ -12,6 +12,7 @@ import { Link } from 'react-router-dom';
 
 import UserItem from './User';
 import { withStyles } from 'material-ui/styles';
+import { Context } from '../../../../App';
 
 // import Modal from './AuthModal';
 
@@ -45,12 +46,13 @@ export class MainMenu extends Component {
   }
 
 
-  static contextTypes = {
-    logout: PropTypes.func.isRequired,
-    onAuthSuccess: PropTypes.func.isRequired,
-    user: PropTypes.object,
-    openLoginForm: PropTypes.func.isRequired,
-  }
+  // static contextTypes = {
+  //   logout: PropTypes.func.isRequired,
+  //   onAuthSuccess: PropTypes.func.isRequired,
+  //   user: PropTypes.object,
+  //   openLoginForm: PropTypes.func.isRequired,
+  // }
+  static contextType = Context;
 
   state = {
     // opened: false,